post 관련 데이터 가져오는 함수들을 용도에 따라 파일로 분리
